Remove nominal typing via abc.ABC
