+++ src/screener/reader/kindle.py
@@ -1,24 +1,22 @@
 from __future__ import annotations
 
 import shutil
-from typing import TYPE_CHECKING
+import typing
 
-if TYPE_CHECKING:
+if typing.TYPE_CHECKING:
     from pathlib import Path
     from types import TracebackType
 
 from mobi import extract
 
-from .abstract import AbstractReader
 
-
-class KindleFileReader(AbstractReader):
-    def __init__(self: KindleFileReader, file_path: Path) -> None:
-        super().__init__(file_path)
+class KindleFileReader:
+    def __init__(self, file_path: Path) -> None:
+        self.file_path = file_path
         self._temp_dir = None
         self.generated_translation = None
 
-    def __enter__(self: KindleFileReader) -> KindleFileReader:
+    def __enter__(self) -> KindleFileReader:
         # The `mobi` library likes file paths as strings for whatever reason.
         self._temp_dir, self.generated_translation = extract(str(self.file_path))
         if self.generated_translation is None or self._temp_dir is None:
@@ -26,8 +24,19 @@ class KindleFileReader(AbstractReader):
             raise ValueError(msg)
         return self
 
+    @typing.overload
+    def __exit__(self, exc_type: None, exc_val: None, exc_tb: None) -> None: ...
+
+    @typing.overload
     def __exit__(
-        self: KindleFileReader,
+        self,
+        exc_type: type[BaseException],
+        exc_val: BaseException,
+        exc_tb: TracebackType,
+    ) -> None: ...
+
+    def __exit__(
+        self,
         exc_type: type[BaseException] | None,
         exc_val: BaseException | None,
         exc_tb: TracebackType | None,
